+++ devand-crypto/src/lib.rs
@@ -1,2 +1,1 @@
-mod password_reset_token;
-mod jwt;
+mod signed_token;
